refactor(examples): extract EngineDeck theme and chart builders
EngineDeckExample carried its 21-colour palette and every benchmark
table/chart builder inline, at 862 lines. Move the palette to
EngineDeckTheme and the table + chart specs to EngineDeckCharts, both
package-private and statically imported — completing the
data/theme/widgets split the deck began with EngineDeckData.

EngineDeckExample drops to 659 lines and now only composes pages. The
layout snapshot stays byte-identical, proving the move changed no output.
